ISSUE #3307 - refactor index files
