+++ frontend/src/v5/ui/components/shared/navigationCard/projectCard/index.ts
@@ -14,5 +14,6 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
-export { PlaceholderCard } from './placeholderCard.component';
+export { ProjectCard } from './projectCard.component';
+export { AddProjectCard } from './addProjectCard/addProjectCard.component';
+export { PlaceholderCard } from './placeholderCard/placeholderCard.component';
